+++ TransitionTreasury/TransitionAnimation/PageTransitionAnimation.swift
@@ -12,7 +12,7 @@ public class PageTransitionAnimation: NSObject, TRViewControllerAnimatedTransiti
     
     public var transitionStatus: TransitionStatus?
     
-    public var previousStatusBarStyle: UIStatusBarStyle?
+    public var previousStatusBarStyle: TRStatusBarStyle?
     
     public var transitionContext: UIViewControllerContextTransitioning?