+++ TransitionTreasury/TransitionAnimation/PopTipTransitionAnimation.swift
@@ -12,7 +12,7 @@ public class PopTipTransitionAnimation: NSObject, TRViewControllerAnimatedTransi
     
     public var transitionStatus: TransitionStatus?
     
-    public var previousStatusBarStyle: UIStatusBarStyle?
+    public var previousStatusBarStyle: TRStatusBarStyle?
     
     public var transitionContext: UIViewControllerContextTransitioning?