+++ TransitionTreasury/TransitionAnimation/TaaskyFlipTransitionAnimation.swift
@@ -12,7 +12,7 @@ public class TaaskyFlipTransitionAnimation: NSObject, TRViewControllerAnimatedTr
     
     public var transitionStatus: TransitionStatus?
     
-    public var previousStatusBarStyle: UIStatusBarStyle?
+    public var previousStatusBarStyle: TRStatusBarStyle?
     
     public var transitionContext: UIViewControllerContextTransitioning?