+++ TransitionTreasury/TransitionAnimation/TwitterTransitionAnimation.swift
@@ -12,7 +12,7 @@ public class TwitterTransitionAnimation: NSObject, TRViewControllerAnimatedTrans
     
     public var transitionStatus: TransitionStatus?
     
-    public var previousStatusBarStyle: UIStatusBarStyle?
+    public var previousStatusBarStyle: TRStatusBarStyle?
     
     public var transitionContext: UIViewControllerContextTransitioning?