+++ CHANGELOG.md
@@ -14,7 +14,15 @@
 * Fade    
 * PopTip   
 * TaaskyFlip    
-* Elevate    
+* Elevate  
+
+## [1.1.1](https://github.com/DianQK/TransitionTreasury/releases/tag/1.0.1)   
+
+#### Fixed   
+* Fix no `Hide` status bar style.   
+
+#### Other
+* Separate status bar style module.
 
 ## [1.1.0](https://github.com/DianQK/TransitionTreasury/releases/tag/1.0.0)